replaced interface location and calls
