+++ src/types/valuation.type.ts
@@ -1,0 +1,5 @@
+import { SingleLandAPIResponse } from "../interfaces/land.interface";
+
+export type LandsKey = 'sandbox' | 'decentraland' | 'axie-infinity' | 'somnium-space';
+
+export type LandListAPIResponse = Record<string, SingleLandAPIResponse>
